vibevirtuoso-database/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2
pymongo[srv]==4.6.0

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Environment and configuration
python-dotenv==1.0.0

# HTTP client (for testing)
httpx==0.25.2